Update File_Monitoring_Legacy.py
This update now checks to see if the files to delete list is empty, if it is it means there are no files to delete and will therefore wait

+++ Misc/Programs/File_Monitoring_Legacy.py
@@ -32,13 +32,19 @@ def get_files_to_delete(path):
 # [3] Deletes the files
 def delete_files(path):
     files_to_delete = get_files_to_delete(path)
+    if len(files_to_delete) == 0: # Checks to see if the list is empty, If it is it means there's no files to delete
+        #print("Skip")
+        sleep(10)
+
+    else:  
     
-    try: 
-        # Deletes the files that are blank
-        for i in range(len(files_to_delete)):
-            os.remove(files_to_delete[i])
-    except:
-        pass
+        try: 
+            # Deletes the files that are blank
+            for i in range(len(files_to_delete)):
+                os.remove(files_to_delete[i])
+                #print(f"Deleted: {files_to_delete[i]}")
+        except:
+            pass
     
 
 # [4] Runs the Program
@@ -59,3 +65,16 @@ def run(path):
 #print(get_files_to_delete(path)) 
 
 #run(path)
+
+"""trading_pairs = ["BTCUSDT"] #, "ETHUSDT"]
+
+path_list = []
+for i in range(len(trading_pairs)):
+    file_path = f"2-DataProcessing/data_gathered/{trading_pairs[i]}_data"
+    
+    path_list.append(file_path)
+
+
+
+
+(run(path_list))"""
